Remove assertion that all indices must be removed
This is not true. e.g:

E --<child>----->B(owned, open)
E --<extension>->A(closed)

We want to keep B. However, we must remove E since it is the extension
of a closed case (A). Since E has a child relationship to B, and B isn't
in the list, this assertion no longer holds